+++ TractorMod/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* When you have no audio device connected, custom tractor sounds are now disabled automatically to avoid an error related to a bug in the base game.
+
 ## 4.24.3
 Released 11 January 2026 for SMAPI 4.1.10 or later.